refactor(platform): Refactor project components (#626)
