+++ apps/api/package.json
@@ -22,6 +22,7 @@
     "unit": "pnpm db:generate-types && jest --config=jest.config.ts"
   },
   "dependencies": {
+    "@keyshade/schema": "workspace:*",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.2.0",
     "@nestjs/core": "^10.0.0",
@@ -34,18 +35,22 @@
     "@nestjs/swagger": "^7.3.0",
     "@nestjs/throttler": "^6.2.1",
     "@nestjs/websockets": "^10.3.7",
-    "@slack/bolt": "^3.22.0",
     "@react-email/components": "^0.0.25",
     "@react-email/preview": "0.0.11",
+    "@slack/bolt": "^3.22.0",
     "@react-email/render": "^1.0.1",
     "@socket.io/redis-adapter": "^8.3.0",
+    "@sentry/node": "^7.102.0",
+    "@sentry/profiling-node": "^7.102.0",
     "@supabase/supabase-js": "^2.39.6",
+    "chalk": "^4.1.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
     "cookie-parser": "^1.4.6",
     "dayjs": "^1.11.11",
     "eccrypto": "^1.1.6",
     "minio": "^8.0.0",
+    "moment": "^2.30.1",
     "nodemailer": "^6.9.9",
     "passport-github2": "^0.1.12",
     "passport-gitlab2": "^5.0.0",
@@ -54,9 +59,11 @@
     "redis": "^4.6.13",
     "rxjs": "^7.8.1",
     "socket.io": "^4.7.5",
-    "uuid": "^9.0.1"
+    "uuid": "^9.0.1",
+    "zod": "^3.23.6"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.2",
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
@@ -69,8 +76,10 @@
     "@types/supertest": "^6.0.0",
     "@types/uuid": "^9.0.8",
     "ajv": "^7",
+    "cross-env": "^7.0.3",
     "dotenv-cli": "^7.4.2",
     "file-loader": "^6.2.0",
+    "husky": "^9.0.11",
     "jest": "^29.5.0",
     "jest-mock-extended": "^3.0.5",
     "prettier": "^3.0.0",
