+++ apps/api/tsconfig.app.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.json",
+  "extends": "tsconfig.json",
   "compilerOptions": {
     "outDir": "dist/out-tsc",
     "module": "ES2022",
