+++ apps/platform/src/store/workspace.ts
